+++ sessions/internal/session/mocks/sessions_usecase_mock.go
@@ -33,7 +33,7 @@ func (m *MockUsecase) EXPECT() *MockUsecaseMockRecorder {
 }
 
 // AddToken mocks base method
-func (m *MockUsecase) AddToken(ID int) (string, error) {
+func (m *MockUsecase) AddToken(ID int64) (string, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "AddToken", ID)
 	ret0, _ := ret[0].(string)
@@ -48,10 +48,10 @@ func (mr *MockUsecaseMockRecorder) AddToken(ID interface{}) *gomock.Call {
 }
 
 // GetIDByToken mocks base method
-func (m *MockUsecase) GetIDByToken(token string) (int, error) {
+func (m *MockUsecase) GetIDByToken(token string) (int64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetIDByToken", token)
-	ret0, _ := ret[0].(int)
+	ret0, _ := ret[0].(int64)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
